+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
-## [Unreleased]
+## [25.2.1] - 2025-11-25
 
 ### Added
 
@@ -16,3 +16,14 @@ and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 - Palettes
 - Color tools
 - Palette draft saving to local storage
+- Palette editablility propagation
+- Toolbar
+- Made palette name editable
+
+### Changed
+
+- `_lsItemPrefix` and `_lsIndexKey` from getters to static methods
+
+### Fixed
+
+- Parsing for `PaletteTheme` from v1 object